+++ spring/build.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
     kotlin("jvm") version "1.9.0"
-
     `maven-publish`
 }
