+++ spring/src/main/kotlin/com/katanox/tabour/spring/config/AutoconfigureTabour.kt
@@ -3,4 +3,4 @@ package com.katanox.tabour.spring.config
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.RUNTIME)
 @MustBeDocumented
-annotation class TabourAutoConfiguration(val enable: Boolean = false)
+annotation class AutoconfigureTabour
